PB/csc309project/requirements.txt
asgiref==3.5.2
cffi==1.15.1
colorama==0.4.6
cryptography==38.0.3
Django==4.0.8
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
geographiclib==2.0
geopy==2.3.0
phonenumbers==8.13.0
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.6
sqlparse==0.4.3
